railway-deployment/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
pandas==2.2.3
python-dotenv==1.0.0

# Core ML
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.5.1
huggingface-hub==0.20.0

# Vector Search
faiss-cpu==1.7.4

# Critical pin: fix ABI errors with faiss
numpy==1.26.4